refactor: move mocks alongside interfaces
